code/requirements.txt
psycopg2==2.9.3
requests==2.27.1
PyYAML==6.0
flask
datetimerange==1.2.0
python-dateutil==2.8.2
phonenumbers==8.12
sqlescapy==1.0.1
fastapi==0.78.0
schedule==1.1
python-jose[cryptography]==3.3
pydantic[email]==1.9.1
passlib[bcrypt]==1.7
python-multipart==0.0.5
uvicorn
email-validator==1.1.2
PyJWT==1.7.1
python-decouple==3.3